chanstate: move open channel read kv store
Move the KV-backed open channel read and scan methods onto KVStore.
ChannelStateDB keeps compatibility wrappers that attach themselves to
returned OpenChannel values so existing receiver methods still use the
channeldb store.

Leave restore and sync paths in channeldb because they still coordinate
link-node creation and repair while LinkNodeDB ownership remains outside
chanstate.
