+++ channeldb/db.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/wire/v2"
-	"github.com/btcsuite/btcwallet/walletdb"
 	mig "github.com/lightningnetwork/lnd/channeldb/migration"
 	"github.com/lightningnetwork/lnd/channeldb/migration12"
 	"github.com/lightningnetwork/lnd/channeldb/migration13"
@@ -580,16 +579,12 @@ func (c *ChannelStateDB) LinkNodeDB() *LinkNodeDB {
 func (c *ChannelStateDB) FetchOpenChannels(nodeID *btcec.PublicKey) (
 	[]*OpenChannel, error) {
 
-	var channels []*OpenChannel
-	err := kvdb.View(c.backend, func(tx kvdb.RTx) error {
-		var err error
-		channels, err = c.fetchOpenChannels(tx, nodeID)
-		return err
-	}, func() {
-		channels = nil
-	})
+	channels, err := c.kvStore.FetchOpenChannels(nodeID)
+	if err != nil {
+		return nil, err
+	}
 
-	return channels, err
+	return c.attachOpenChannelStores(channels), nil
 }
 
 // fetchOpenChannels uses and existing database transaction and returns all
@@ -599,115 +594,32 @@ func (c *ChannelStateDB) FetchOpenChannels(nodeID *btcec.PublicKey) (
 func (c *ChannelStateDB) fetchOpenChannels(tx kvdb.RTx,
 	nodeID *btcec.PublicKey) ([]*OpenChannel, error) {
 
-	// Get the bucket dedicated to storing the metadata for open channels.
-	openChanBucket := tx.ReadBucket(openChannelBucket)
-	if openChanBucket == nil {
-		return nil, nil
-	}
-
-	// Within this top level bucket, fetch the bucket dedicated to storing
-	// open channel data specific to the remote node.
-	pub := nodeID.SerializeCompressed()
-	nodeChanBucket := openChanBucket.NestedReadBucket(pub)
-	if nodeChanBucket == nil {
-		return nil, nil
-	}
-
-	// Next, we'll need to go down an additional layer in order to retrieve
-	// the channels for each chain the node knows of.
-	var channels []*OpenChannel
-	err := nodeChanBucket.ForEach(func(chainHash, v []byte) error {
-		// If there's a value, it's not a bucket so ignore it.
-		if v != nil {
-			return nil
-		}
-
-		// If we've found a valid chainhash bucket, then we'll retrieve
-		// that so we can extract all the channels.
-		chainBucket := nodeChanBucket.NestedReadBucket(chainHash)
-		if chainBucket == nil {
-			return fmt.Errorf("unable to read bucket for chain=%x",
-				chainHash[:])
-		}
-
-		// Finally, we both of the necessary buckets retrieved, fetch
-		// all the active channels related to this node.
-		nodeChannels, err := c.fetchNodeChannels(tx, chainBucket)
-		if err != nil {
-			return fmt.Errorf("unable to read channel for "+
-				"chain_hash=%x, node_key=%x: %v",
-				chainHash[:], pub, err)
-		}
-
-		channels = append(channels, nodeChannels...)
-		return nil
-	})
-
-	return channels, err
-}
-
-// fetchNodeChannels retrieves all active channels from the target chainBucket
-// which is under a node's dedicated channel bucket. This function is typically
-// used to fetch all the active channels related to a particular node. Channels
-// already flipped to outpointClosed in the outpoint index are skipped silently
-// — readers see only channels that are still considered open.
-func (c *ChannelStateDB) fetchNodeChannels(tx kvdb.RTx,
-	chainBucket kvdb.RBucket) ([]*OpenChannel, error) {
-
-	var channels []*OpenChannel
-
-	// Hoist the outpoint-bucket lookup so the closed-channel check inside
-	// the loop is a per-iteration map probe rather than a tx-level bucket
-	// resolve.
-	opBucket := tx.ReadBucket(outpointBucket)
-
-	// A node may have channels on several chains, so for each known chain,
-	// we'll extract all the channels.
-	err := chainBucket.ForEach(func(chanPoint, v []byte) error {
-		// If there's a value, it's not a bucket so ignore it.
-		if v != nil {
-			return nil
-		}
-
-		// Skip already-closed channels. The chanBucket still exists
-		// on disk on tombstone-enabled backends; the outpoint flip is
-		// the sole signal that the channel should be treated as
-		// closed.
-		isClosed, err := isOutpointClosed(opBucket, chanPoint)
-		if err != nil {
-			return err
-		}
-		if isClosed {
-			return nil
-		}
-
-		// Once we've found a valid channel bucket, we'll extract it
-		// from the node's chain bucket.
-		chanBucket := chainBucket.NestedReadBucket(chanPoint)
-
-		var outPoint wire.OutPoint
-		err = graphdb.ReadOutpoint(
-			bytes.NewReader(chanPoint), &outPoint,
-		)
-		if err != nil {
-			return err
-		}
-		oChannel, err := fetchOpenChannel(chanBucket, &outPoint)
-		if err != nil {
-			return fmt.Errorf("unable to read channel data for "+
-				"chan_point=%v: %w", outPoint, err)
-		}
-		oChannel.Db = c
-
-		channels = append(channels, oChannel)
-
-		return nil
-	})
+	channels, err := chanstate.FetchOpenChannelsTx(tx, nodeID)
 	if err != nil {
 		return nil, err
 	}
 
-	return channels, nil
+	return c.attachOpenChannelStores(channels), nil
+}
+
+func (c *ChannelStateDB) attachOpenChannelStore(
+	channel *OpenChannel) *OpenChannel {
+
+	if channel != nil {
+		channel.Db = c
+	}
+
+	return channel
+}
+
+func (c *ChannelStateDB) attachOpenChannelStores(
+	channels []*OpenChannel) []*OpenChannel {
+
+	for _, channel := range channels {
+		c.attachOpenChannelStore(channel)
+	}
+
+	return channels
 }
 
 // FetchChannel attempts to locate a channel specified by the passed channel
@@ -715,20 +627,12 @@ func (c *ChannelStateDB) fetchNodeChannels(tx kvdb.RTx,
 func (c *ChannelStateDB) FetchChannel(chanPoint wire.OutPoint) (*OpenChannel,
 	error) {
 
-	var targetChanPoint bytes.Buffer
-	err := graphdb.WriteOutpoint(&targetChanPoint, &chanPoint)
+	channel, err := c.kvStore.FetchChannel(chanPoint)
 	if err != nil {
 		return nil, err
 	}
 
-	targetChanPointBytes := targetChanPoint.Bytes()
-	selector := func(chainBkt walletdb.ReadBucket) ([]byte, *wire.OutPoint,
-		error) {
-
-		return targetChanPointBytes, &chanPoint, nil
-	}
-
-	return c.channelScanner(nil, selector)
+	return c.attachOpenChannelStore(channel), nil
 }
 
 // FetchChannelByID attempts to locate a channel specified by the passed channel
@@ -736,48 +640,12 @@ func (c *ChannelStateDB) FetchChannel(chanPoint wire.OutPoint) (*OpenChannel,
 func (c *ChannelStateDB) FetchChannelByID(id lnwire.ChannelID) (*OpenChannel,
 	error) {
 
-	selector := func(chainBkt walletdb.ReadBucket) ([]byte, *wire.OutPoint,
-		error) {
-
-		var (
-			targetChanPointBytes []byte
-			targetChanPoint      *wire.OutPoint
-
-			// errChanFound is used to signal that the channel has
-			// been found so that iteration through the DB buckets
-			// can stop.
-			errChanFound = errors.New("channel found")
-		)
-		err := chainBkt.ForEach(func(k, _ []byte) error {
-			var outPoint wire.OutPoint
-			err := graphdb.ReadOutpoint(
-				bytes.NewReader(k), &outPoint,
-			)
-			if err != nil {
-				return err
-			}
-
-			chanID := lnwire.NewChanIDFromOutPoint(outPoint)
-			if chanID != id {
-				return nil
-			}
-
-			targetChanPoint = &outPoint
-			targetChanPointBytes = k
-
-			return errChanFound
-		})
-		if err != nil && !errors.Is(err, errChanFound) {
-			return nil, nil, err
-		}
-		if targetChanPoint == nil {
-			return nil, nil, ErrChannelNotFound
-		}
-
-		return targetChanPointBytes, targetChanPoint, nil
+	channel, err := c.kvStore.FetchChannelByID(id)
+	if err != nil {
+		return nil, err
 	}
 
-	return c.channelScanner(nil, selector)
+	return c.attachOpenChannelStore(channel), nil
 }
 
 // ChanCount is used by the server in determining access control.
@@ -789,373 +657,43 @@ type ChanCount = chanstate.ChanCount
 func (c *ChannelStateDB) FetchPermAndTempPeers(
 	chainHash []byte) (map[string]ChanCount, error) {
 
-	peerChanInfo := make(map[string]ChanCount)
-
-	err := kvdb.View(c.backend, func(tx kvdb.RTx) error {
-		openChanBucket := tx.ReadBucket(openChannelBucket)
-		if openChanBucket == nil {
-			return ErrNoChanDBExists
-		}
-
-		// Hoist the outpoint-bucket lookup so the closed-channel check
-		// inside the nested chainBucket.ForEach below is a per-channel
-		// map probe rather than a tx-level bucket resolve.
-		opBucket := tx.ReadBucket(outpointBucket)
-
-		openChanErr := openChanBucket.ForEach(func(nodePub,
-			v []byte) error {
-
-			// If there is a value, this is not a bucket.
-			if v != nil {
-				return nil
-			}
-
-			nodeChanBucket := openChanBucket.NestedReadBucket(
-				nodePub,
-			)
-			if nodeChanBucket == nil {
-				return nil
-			}
-
-			chainBucket := nodeChanBucket.NestedReadBucket(
-				chainHash,
-			)
-			if chainBucket == nil {
-				return fmt.Errorf("no chain bucket exists")
-			}
-
-			var isPermPeer bool
-			var pendingOpenCount uint64
-
-			internalErr := chainBucket.ForEach(func(chanPoint,
-				val []byte) error {
-
-				// If there is a value, this is not a bucket.
-				if val != nil {
-					return nil
-				}
-
-				// Skip already-closed channels: they are
-				// logically closed even though their
-				// per-channel state still resides under
-				// chainBucket. The closed peer's protected
-				// status is established below via the
-				// historical-channel scan.
-				isClosed, err := isOutpointClosed(
-					opBucket, chanPoint,
-				)
-				if err != nil {
-					return err
-				}
-				if isClosed {
-					return nil
-				}
-
-				chanBucket := chainBucket.NestedReadBucket(
-					chanPoint,
-				)
-				if chanBucket == nil {
-					return nil
-				}
-
-				var op wire.OutPoint
-				readErr := graphdb.ReadOutpoint(
-					bytes.NewReader(chanPoint), &op,
-				)
-				if readErr != nil {
-					return readErr
-				}
-
-				// We need to go through each channel and look
-				// at the IsPending status.
-				openChan, err := fetchOpenChannel(
-					chanBucket, &op,
-				)
-				if err != nil {
-					return err
-				}
-
-				if openChan.IsPending {
-					// Add to the pending-open count since
-					// this is a temp peer.
-					pendingOpenCount++
-					return nil
-				}
-
-				// Since IsPending is false, this is a perm
-				// peer.
-				isPermPeer = true
-
-				return nil
-			})
-			if internalErr != nil {
-				return internalErr
-			}
-
-			peerCount := ChanCount{
-				HasOpenOrClosedChan: isPermPeer,
-				PendingOpenCount:    pendingOpenCount,
-			}
-			peerChanInfo[string(nodePub)] = peerCount
-
-			return nil
-		})
-		if openChanErr != nil {
-			return openChanErr
-		}
-
-		// Now check the closed channel bucket.
-		historicalChanBucket := tx.ReadBucket(historicalChannelBucket)
-		if historicalChanBucket == nil {
-			return ErrNoHistoricalBucket
-		}
-
-		historicalErr := historicalChanBucket.ForEach(func(chanPoint,
-			v []byte) error {
-			// Parse each nested bucket and the chanInfoKey to get
-			// the IsPending bool. This determines whether the
-			// peer is protected or not.
-			if v != nil {
-				// This is not a bucket. This is currently not
-				// possible.
-				return nil
-			}
-
-			chanBucket := historicalChanBucket.NestedReadBucket(
-				chanPoint,
-			)
-			if chanBucket == nil {
-				// This is not possible.
-				return fmt.Errorf("no historical channel " +
-					"bucket exists")
-			}
-
-			var op wire.OutPoint
-			readErr := graphdb.ReadOutpoint(
-				bytes.NewReader(chanPoint), &op,
-			)
-			if readErr != nil {
-				return readErr
-			}
-
-			// This channel is closed, but the structure of the
-			// historical bucket is the same. This is by design,
-			// which means we can call fetchOpenChannel.
-			channel, fetchErr := fetchOpenChannel(chanBucket, &op)
-			if fetchErr != nil {
-				return fetchErr
-			}
-
-			// Only include this peer in the protected class if
-			// the closing transaction confirmed. Note that
-			// CloseChannel can be called in the funding manager
-			// while IsPending is true which is why we need this
-			// special-casing to not count premature funding
-			// manager calls to CloseChannel.
-			if !channel.IsPending {
-				// Fetch the public key of the remote node. We
-				// need to use the string-ified serialized,
-				// compressed bytes as the key.
-				remotePub := channel.IdentityPub
-				remoteSer := remotePub.SerializeCompressed()
-				remoteKey := string(remoteSer)
-
-				count, exists := peerChanInfo[remoteKey]
-				if exists {
-					count.HasOpenOrClosedChan = true
-					peerChanInfo[remoteKey] = count
-				} else {
-					peerCount := ChanCount{
-						HasOpenOrClosedChan: true,
-					}
-					peerChanInfo[remoteKey] = peerCount
-				}
-			}
-
-			return nil
-		})
-		if historicalErr != nil {
-			return historicalErr
-		}
-
-		return nil
-	}, func() {
-		clear(peerChanInfo)
-	})
-
-	return peerChanInfo, err
-}
-
-// channelSelector describes a function that takes a chain-hash bucket from
-// within the open-channel DB and returns the wanted channel point bytes, and
-// channel point. It must return the ErrChannelNotFound error if the wanted
-// channel is not in the given bucket.
-type channelSelector func(chainBkt walletdb.ReadBucket) ([]byte, *wire.OutPoint,
-	error)
-
-// channelScanner will traverse the DB to each chain-hash bucket of each node
-// pub-key bucket in the open-channel-bucket. The chanSelector will then be used
-// to fetch the wanted channel outpoint from the chain bucket.
-func (c *ChannelStateDB) channelScanner(tx kvdb.RTx,
-	chanSelect channelSelector) (*OpenChannel, error) {
-
-	var (
-		targetChan *OpenChannel
-
-		// errChanFound is used to signal that the channel has been
-		// found so that iteration through the DB buckets can stop.
-		errChanFound = errors.New("channel found")
-	)
-
-	// chanScan will traverse the following bucket structure:
-	//  * nodePub => chainHash => chanPoint
-	//
-	// At each level we go one further, ensuring that we're traversing the
-	// proper key (that's actually a bucket). By only reading the bucket
-	// structure and skipping fully decoding each channel, we save a good
-	// bit of CPU as we don't need to do things like decompress public
-	// keys.
-	chanScan := func(tx kvdb.RTx) error {
-		// Get the bucket dedicated to storing the metadata for open
-		// channels.
-		openChanBucket := tx.ReadBucket(openChannelBucket)
-		if openChanBucket == nil {
-			return ErrNoActiveChannels
-		}
-
-		// Hoist the outpoint-bucket lookup so the closed-channel
-		// check inside the per-chain ForEach below pays one tx-level
-		// bucket resolve total instead of one per visited chanKey.
-		opBucket := tx.ReadBucket(outpointBucket)
-
-		// Within the node channel bucket, are the set of node pubkeys
-		// we have channels with, we don't know the entire set, so we'll
-		// check them all.
-		return openChanBucket.ForEach(func(nodePub, v []byte) error {
-			// Ensure that this is a key the same size as a pubkey,
-			// and also that it leads directly to a bucket.
-			if len(nodePub) != 33 || v != nil {
-				return nil
-			}
-
-			nodeChanBucket := openChanBucket.NestedReadBucket(
-				nodePub,
-			)
-			if nodeChanBucket == nil {
-				return nil
-			}
-
-			// The next layer down is all the chains that this node
-			// has channels on with us.
-			return nodeChanBucket.ForEach(func(chainHash,
-				v []byte) error {
-
-				// If there's a value, it's not a bucket so
-				// ignore it.
-				if v != nil {
-					return nil
-				}
-
-				chainBucket := nodeChanBucket.NestedReadBucket(
-					chainHash,
-				)
-				if chainBucket == nil {
-					return fmt.Errorf("unable to read "+
-						"bucket for chain=%x",
-						chainHash)
-				}
-
-				// Finally, we reach the leaf bucket that stores
-				// all the chanPoints for this node.
-				targetChanBytes, chanPoint, err := chanSelect(
-					chainBucket,
-				)
-				if errors.Is(err, ErrChannelNotFound) {
-					return nil
-				} else if err != nil {
-					return err
-				}
-
-				// An already-closed channel is logically gone
-				// and must not be surfaced by lookup-style
-				// scans.
-				isClosed, err := isOutpointClosed(
-					opBucket, targetChanBytes,
-				)
-				if err != nil {
-					return err
-				}
-				if isClosed {
-					return nil
-				}
-
-				chanBucket := chainBucket.NestedReadBucket(
-					targetChanBytes,
-				)
-				if chanBucket == nil {
-					return nil
-				}
-
-				channel, err := fetchOpenChannel(
-					chanBucket, chanPoint,
-				)
-				if err != nil {
-					return err
-				}
-
-				targetChan = channel
-				targetChan.Db = c
-
-				return errChanFound
-			})
-		})
-	}
-
-	var err error
-	if tx == nil {
-		err = kvdb.View(c.backend, chanScan, func() {})
-	} else {
-		err = chanScan(tx)
-	}
-	if err != nil && !errors.Is(err, errChanFound) {
-		return nil, err
-	}
-
-	if targetChan != nil {
-		return targetChan, nil
-	}
-
-	// If we can't find the channel, then we return with an error, as we
-	// have nothing to back up.
-	return nil, ErrChannelNotFound
+	return c.kvStore.FetchPermAndTempPeers(chainHash)
 }
 
 // FetchAllChannels attempts to retrieve all open channels currently stored
 // within the database, including pending open, fully open and channels waiting
 // for a closing transaction to confirm.
 func (c *ChannelStateDB) FetchAllChannels() ([]*OpenChannel, error) {
-	return fetchChannels(c)
+	channels, err := c.kvStore.FetchAllChannels()
+	if err != nil {
+		return nil, err
+	}
+
+	return c.attachOpenChannelStores(channels), nil
 }
 
 // FetchAllOpenChannels will return all channels that have the funding
 // transaction confirmed, and is not waiting for a closing transaction to be
 // confirmed.
 func (c *ChannelStateDB) FetchAllOpenChannels() ([]*OpenChannel, error) {
-	return fetchChannels(
-		c,
-		pendingChannelFilter(false),
-		waitingCloseFilter(false),
-	)
+	channels, err := c.kvStore.FetchAllOpenChannels()
+	if err != nil {
+		return nil, err
+	}
+
+	return c.attachOpenChannelStores(channels), nil
 }
 
 // FetchPendingChannels will return channels that have completed the process of
 // generating and broadcasting funding transactions, but whose funding
 // transactions have yet to be confirmed on the blockchain.
 func (c *ChannelStateDB) FetchPendingChannels() ([]*OpenChannel, error) {
-	return fetchChannels(c,
-		pendingChannelFilter(true),
-		waitingCloseFilter(false),
-	)
+	channels, err := c.kvStore.FetchPendingChannels()
+	if err != nil {
+		return nil, err
+	}
+
+	return c.attachOpenChannelStores(channels), nil
 }
 
 // FetchWaitingCloseChannels will return all channels that have been opened,
@@ -1163,138 +701,12 @@ func (c *ChannelStateDB) FetchPendingChannels() ([]*OpenChannel, error) {
 //
 // NOTE: This includes channels that are also pending to be opened.
 func (c *ChannelStateDB) FetchWaitingCloseChannels() ([]*OpenChannel, error) {
-	return fetchChannels(
-		c, waitingCloseFilter(true),
-	)
-}
-
-// fetchChannelsFilter applies a filter to channels retrieved in fetchchannels.
-// A set of filters can be combined to filter across multiple dimensions.
-type fetchChannelsFilter func(channel *OpenChannel) bool
-
-// pendingChannelFilter returns a filter based on whether channels are pending
-// (ie, their funding transaction still needs to confirm). If pending is false,
-// channels with confirmed funding transactions are returned.
-func pendingChannelFilter(pending bool) fetchChannelsFilter {
-	return func(channel *OpenChannel) bool {
-		return channel.IsPending == pending
-	}
-}
-
-// waitingCloseFilter returns a filter which filters channels based on whether
-// they are awaiting the confirmation of their closing transaction. If waiting
-// close is true, channels that have had their closing tx broadcast are
-// included. If it is false, channels that are not awaiting confirmation of
-// their close transaction are returned.
-func waitingCloseFilter(waitingClose bool) fetchChannelsFilter {
-	return func(channel *OpenChannel) bool {
-		// If the channel is in any other state than Default,
-		// then it means it is waiting to be closed.
-		channelWaitingClose :=
-			channel.ChanStatus() != ChanStatusDefault
-
-		// Include the channel if it matches the value for
-		// waiting close that we are filtering on.
-		return channelWaitingClose == waitingClose
-	}
-}
-
-// fetchChannels attempts to retrieve channels currently stored in the
-// database. It takes a set of filters which are applied to each channel to
-// obtain a set of channels with the desired set of properties. Only channels
-// which have a true value returned for *all* of the filters will be returned.
-// If no filters are provided, every channel in the open channels bucket will
-// be returned.
-func fetchChannels(c *ChannelStateDB, filters ...fetchChannelsFilter) (
-	[]*OpenChannel, error) {
-
-	var channels []*OpenChannel
-
-	err := kvdb.View(c.backend, func(tx kvdb.RTx) error {
-		// Get the bucket dedicated to storing the metadata for open
-		// channels.
-		openChanBucket := tx.ReadBucket(openChannelBucket)
-		if openChanBucket == nil {
-			return ErrNoActiveChannels
-		}
-
-		// Next, fetch the bucket dedicated to storing metadata related
-		// to all nodes. All keys within this bucket are the serialized
-		// public keys of all our direct counterparties.
-		nodeMetaBucket := tx.ReadBucket(nodeInfoBucket)
-		if nodeMetaBucket == nil {
-			return fmt.Errorf("node bucket not created")
-		}
-
-		// Finally for each node public key in the bucket, fetch all
-		// the channels related to this particular node.
-		return nodeMetaBucket.ForEach(func(k, v []byte) error {
-			nodeChanBucket := openChanBucket.NestedReadBucket(k)
-			if nodeChanBucket == nil {
-				return nil
-			}
-
-			return nodeChanBucket.ForEach(func(chainHash, v []byte) error {
-				// If there's a value, it's not a bucket so
-				// ignore it.
-				if v != nil {
-					return nil
-				}
-
-				// If we've found a valid chainhash bucket,
-				// then we'll retrieve that so we can extract
-				// all the channels.
-				chainBucket := nodeChanBucket.NestedReadBucket(
-					chainHash,
-				)
-				if chainBucket == nil {
-					return fmt.Errorf("unable to read "+
-						"bucket for chain=%x", chainHash[:])
-				}
-
-				nodeChans, err := c.fetchNodeChannels(
-					tx, chainBucket,
-				)
-				if err != nil {
-					return fmt.Errorf("unable to read "+
-						"channel for chain_hash=%x, "+
-						"node_key=%x: %v", chainHash[:], k, err)
-				}
-				for _, channel := range nodeChans {
-					// includeChannel indicates whether the channel
-					// meets the criteria specified by our filters.
-					includeChannel := true
-
-					// Run through each filter and check whether the
-					// channel should be included.
-					for _, f := range filters {
-						// If the channel fails the filter, set
-						// includeChannel to false and don't bother
-						// checking the remaining filters.
-						if !f(channel) {
-							includeChannel = false
-							break
-						}
-					}
-
-					// If the channel passed every filter, include it in
-					// our set of channels.
-					if includeChannel {
-						channels = append(channels, channel)
-					}
-				}
-				return nil
-			})
-
-		})
-	}, func() {
-		channels = nil
-	})
+	channels, err := c.kvStore.FetchWaitingCloseChannels()
 	if err != nil {
 		return nil, err
 	}
 
-	return channels, nil
+	return c.attachOpenChannelStores(channels), nil
 }
 
 // FetchClosedChannels attempts to fetch all closed channels from the database.
